Update casacore recipe
- Switch to git to pull source code
- Add latest tag
- Fix removal of redundant cmake file to use system one

+++ packages/casacore/package.py
@@ -12,21 +12,22 @@ class Casacore(CMakePackage):
     """A suite of c++ libraries for radio astronomy data processing."""
 
     homepage = "https://github.com/casacore/casacore"
-    url = "https://github.com/casacore/casacore/archive/refs/tags/v3.6.1.tar.gz"
+    git = "https://github.com/casacore/casacore.git"
 
     # TODO: support multiple spack versions (maintainers option)
     # spack version > 18.1 changed the maintainers syntax
     # maintainers("mpokorny")
-    version("3.6.1", sha256="480d3340fa17e9ba67f18efbaff4bbb272a01d1f400d2295c0b6c86eb7abcf82")
-    version("3.5.0", sha256="63f1c8eff932b0fcbd38c598a5811e6e5397b72835b637d6f426105a183b3f91")
-    version("3.4.0", sha256="31f02ad2e26f29bab4a47a2a69e049d7bc511084a0b8263360e6157356f92ae1")
-    version("3.3.0", sha256="3a714644b908ef6e81489b792cc9b80f6d8267a275e15d38a42a6a5137d39d3d")
-    version("3.2.0", sha256="ae5d3786cb6dfdd7ebc5eecc0c724ff02bbf6929720bc23be43a027978e79a5f")
-    version("3.1.2", sha256="ac94f4246412eb45d503f1019cabe2bb04e3861e1f3254b832d9b1164ea5f281")
-    version("3.1.1", sha256="85d2b17d856592fb206b17e0a344a29330650a4269c80b87f8abb3eaf3dadad4")
-    version("3.1.0", sha256="a6adf2d77ad0d6f32995b1e297fd88d31ded9c3e0bb8f28966d7b35a969f7897")
-    version("3.0.0", sha256="6f0e68fd77b5c96299f7583a03a53a90980ec347bff9dfb4c0abb0e2933e6bcb")
-    version("2.4.1", sha256="58eccc875053b2c6fe44fe53b6463030ef169597ec29926936f18d27b5087d63")
+    version("3.7.1")
+    version("3.6.1")
+    version("3.5.0")
+    version("3.4.0")
+    version("3.3.0")
+    version("3.2.0")
+    version("3.1.2")
+    version("3.1.1")
+    version("3.1.0")
+    version("3.0.0")
+    version("2.4.1")
 
     depends_on("cmake@3.7.1:", type="build")
 
@@ -122,4 +123,5 @@ class Casacore(CMakePackage):
             os.system(f"mkdir -p {self.prefix}/share/casacore/data/")
             os.system(f"tar xfz {self.prefix}/WSRT_Measures.ztar -C {self.prefix}/share/casacore/data/")
         # Rely on CMake ability to find hdf5, available since CMake 3.7.X
-        os.remove("cmake/FindHDF5.cmake")
+        if os.path.exists("cmake/FindHDF5.cmake"):
+            os.remove("cmake/FindHDF5.cmake")
